multiple and divide

+++ src/utils.rs
@@ -1,7 +1,7 @@
-use std::f32::EPSILON;
+use std::f64::EPSILON;
 
 /// Compare two floats for equality
-pub fn is_eq_float(x: &f32, y: &f32) -> bool {
+pub fn is_eq_float(x: &f64, y: &f64) -> bool {
     if (x - y).abs() < EPSILON {
         return true;
     }
